HdrHistogram written by Gil Tene can now be used to measure latency instead of our home-grown LinearHistogram.
When a type of a probe is set to "hdr" in a test.property file then a HdrHistogram-based probe will be injected.

Example:
$ cat test.properties
class=com.hazelcast.stabilizer.tests.map.StringMapTest
threadCount=1
keyLocality=remote
probe-putLatency=latency
probe-getLatency=hdr

In this example the getLatency probe will use the HdrHistogram probe. putLatency will use the original LinearHistogram probe.
In the future we can decide to remove the LinearHistogram altogether and make "latency" to be a synonymous to "hdr"

Serialization to XML is still a bit crude, based on base64 encoded internal state of HdrHistogram, however Visualized is able to read & visualize the data.