Improved OperationSelector performance by creating an OperationSelectorBuilder which constructs an OperationSelector instance with fast data structures.
Added an exception if the probability sum of all operations is not exactly 1.0.

Renamed empty() method to addDefaultOperation() and removed method addOperationRemainingProbability() since it has the same effect as adding a default method.

To prevent contention on the random generator a new instance is created for each Worker in all tests which use OperationSelector.