The probe.getIntervalCountAndReset is replaced by a get
Resetting is not a concern of the probe. If a delta needs to be stored
compared to the previous run, it needs to be done somewhere else.

The driver behind this change is that we want to create thread unsafe
probes for higher performance. And resetting from an arbitrary thread
is undesirable.