Control of artifact outut directory.
All artifacts are stored in a directory <testsuite> which is by default
some kind of timestamp, but can be controlled by setting the testsuite id.

This prevent poluting the working directory with a lot of files.

Also the remote workers are automatically downloaded into this directory
and no explicit provisioner --download is needed anymore.