Simplifying HDR usage
This PR simplifes HDR usage so that the HDR probe creates the hdr files. But the system doesn't
know more than that.

Anything that needs to be merged, is triggered from the after-completion.script. Also the hdr
files will automatically be downloaded (and the merged ones creates) due to this after-completion
script and the automatic downloading.

This simplifies the core since it is unaware of hdr.