WorkerType/WorkerParameter improvements
The WorkerParameter is HZ agnostic. The CoordinatorCLI takes care of all the hz magic and there
should not be much remaining HZ logic in the simulator core.

Also there is now a WorkerParameters per workertype.

And there is selection o the right worker startup script based on vendor/worker-type. Which makes it
a lot easier to integrate native clients