DeploymentPlan pick smallest agent
Before this fix the DeploymentPlan starts with the smallest agent (good),
but then moves to the next agent to assign workers, no matter if that
agent is the one with the least amount of workers.

This fix makes sure that for every worker being assigned, always the agent
with the lowest numbers of workers is being picked.

This didn't use to be an issue since we didn't stop/start members, but
with the remote this will happen.