Fixes problem that testcase not stopping
fix #1123

It has been improved so that the worker isn't deleted and follows the regular process
of termination in case of failure. But if a worker has been marked as 'ignore failure'
on the agent, the failure isn't published to the coordinator.

This means that its connections and all that are closed when the worker is terminated
and no communication with a dead worker is scheduled.