Fixes async testing
The problem with async testing is that the timestep method doesn't know the completion,
and therefor can't record the latency. To solve this problem, the timestep method gets
the starttime injected as argument and it is up to the completion handler to record
the right value in the probe.

Fix #1235