Fixes problem with coordinator-remote not completing
The problem is that the call id can't be negative, but in 50%
of the times it was. If the call id is negative, then there is
no registered future and therefor a response doesn't trigger
the future to complete.

Fixes #1268

Adding scripts for running tests

Enhancing scripts for running tests

Commiting changes with LastClusterSizeChange