Simplified warmup
Instead of having a seperate warmup phase (like run), warmup is implemented
by not tracking any performance related information in the warmup period of
the run phase.

So if you have a run duration of 5 minutes and a warmup of 30s, then there
will be performance information for the last 4.5 minute.