It isn't needed to create simulator properties
The variables can be defined and exported in the bash script and will automatically
be picked up without needing to create a file.

Also there is no more need to create an ec2.identity and ec2.credential file. If
the properties refer to a file; the file is loaded. Otherwise the properties will
be interpreted as a direct value.