Ripped out JClouds.
It was a never ending story of things being broken. It jclouds
being tested at all on e.g. ec2??

For introdce a lot more reliability, we just switch to command
line tools of the provider.

So we sacarifice multi cloud stuff for the sake of a working
and controllable platform. What is the point for multicloud support
if making use of a single cloud already is such a big PITA.

Bye bye jclouds disaster!