Fixes orphan process problem
Agents/workers can keep dangling if the coordinator process is killed. This is
especially a problem if the process is run from the IDE.

To solve this problem the agent had the pid of the parent process (for local
setups) and the workers the pid of their agent. If they see that the parent process
is offline by using this parent pid, they terminate themselves.