Fixes problem IBM 6/7
The problem is a bug in the IBM JVM that isn't too happy
when a method exists, that isn't called, in a class. And this
method is pointing to classes that don't exist.

Fix: the explicit casting to sun.management.VMManagement has been
removed. Fix has been verified with IBM 6.