Fixed problem scaling down
The issue was that private ip addresses were used and when there are
multiple vpc of a single user in the same region, there is no guarantee
that ip addresses are unique.

So sometimes based on private ip's the wrong machine id's were found.

This is fixed by making use of public ip address for lookup.