Introduce member-side data ingestion
Normally, entries are ingested from clients. This method can be very
costly if huge amount of entries need to be inserted, because there are
many parties involved in the ingestion: the clients, the network, I/O
and operation threads of the members. This PR intends to shorten this
path by ingesting local data directly on the members.

Along with this mechanism, a modified version of LongByteArrayMapTest is
added: LongByteArrayMapMemberIngestionTest. The difference is that -
since the ingestion is done on the member side and the benchmark phase is
on the client side - the values are not shared. In its current version,
IMapIngestionTask ignores the valueCount and fills the map with random
values.