Version spec from coordinator
The shared simulator.properties can sometimes be a problem if you need to
test with different settings e.g. version_spec.

So the version spec can now be configured fromt the coordinator.

--version maven=4.0.1

This feature is backwards compatible, so no compatibility problems.

Also added some additional examples to the JDK urls