Add template for SQL over TStore and additional tests [HZ-1613]
This will allow the creation of a simulator test dir with TStore and a
storage-optimized EC2 instance already configured. It also includes
tests for mixed SELECT/UPDATE workloads.

Also, this includes some tests for comparison between SQL and
Predicates API performance after the addition of partition-aware client
changes in SQL.