Perftest run auto_gen_report flag addition & some flag fixes
This commit adds the `--auto_gen_report` flag (default true) that allows the skipping of the `perftest report` execution after tests have been run if set to `False`. This is particularly useful for situations such as release verification that are focused on pass/fail instead of performance analysis.

This commit also fixes an issue where some flags were only added to the coordinator arguments list if set to `True`, not when set to `False`. Plus some minor typo fixes.