Added support for automatically commit
When the -c or --commit option is added to perftest run, the changes
of the test git repo will automatically be committed.

When collecting the results, if it is detected that perftest run is running
inside a git repo, the last commit is added to the 'test_commit' file.

This way you can always see the files at the moment the test was run.