Added the AsyncLongStringMapTest
The existing LongStringMapTest (and many others) are not suitable
for generating high levels of concurrency because threads are used
for that. So for a concurrency of 1000, you would need 1000 threads.

The AsyncLongStringMapTest doesn't rely on threads to create concurrency
and therefor can reach migher higher levels of concurrency without the
drawbacks of threads.