Add standalone client tests [PA-547] (#2232)
This PR includes changes to make the new Java Client standalone to be
used by the simulator tests.

It introduces a new driver type: `standaloneclient4plus` You use this
driver if you want to run a client test with the standalone client. The
PR also includes additional changes to make the new driver work. The
driver itself is added in
https://github.com/hazelcast/hazelcast-simulator-drivers/pull/18 as
discussed in the ADR-00045 decider meeting.

Report generation fix:

- Skip 6 lines instead of 5 and rename "buf" column to "buff" in the
dstat parser logic. Dstat output is probably changed in later versions.
The AMI we are using produces a dstat csv file that wasn't working.