Improve dataframe concatenation speed by using more memory [DEX-385] (#2267)
When aggregating all run data the dataframe concatenation is currently
very slow for a run with many workers. The current implementation is
repeatedly copying the same data. This change increases the memory used
by less than the size of the final dataframe but the time taken to run
the operation is greatly improved. For example on my machine the analyze
step takes roughly 2 hours with 6k workers before this change, with this
change it takes less than a minute.