Add capability to define  usage of non-random keys (e.g. to generate near cache hits) [DEX-408] (#2272)
We like to be able to hit to the same keys on a certain probability to
generate some hit ratio for common used keys. This is especially useful
if you like to create some hit ratio for near cache tests.

This change gives the capability to create a hit ratio assuming that the near cache local keys stored
these keys and they are not evicted.

Added 2 new optional test parameters, the `fixedKeyDomain` and
`fixedKeyProbability` to the test. Definition of these fields are:

```
    /**
     * The fixed keys to be used in the test. If set to 0, all
     * keys are random. It should be less than the {@code keyDomain}.
     */
    public int fixedKeyDomain = 0;
    /**
     * The probability of using a fixed key from the {@code fixedKeyDomain}.
     * It is used if {@code fixedKeyPercentage} is set to a value greater than 0.
     */
    public int fixedKeyProbability = 0;
```