Changed the test so that the fixed key set is only used for get operation [DEX-418] (#2278)
Changed the test so that the fixed key set is only used for get
operation to get better deterministic near cache hit ratios.

If you use it also for put operations, it will cause invalidation of the
fixed key which we want to hit and hence decrease desired hit ratio.